requirements.txt
scikit-learn
pandas
numpy
matplotlib
scipy
kmodes
PyYAML
streamlit
